+++ packages/windie-sdk-js/cjs/transport/BackendSocketFactory.js
@@ -3,7 +3,8 @@
  * Provides the backend socket factory module for the TypeScript SDK runtime.
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.createWindieSdkBackendSocket = exports.createAgentBackendSocket = void 0;
+exports.createWindieSdkBackendSocket = void 0;
+exports.createAgentBackendSocket = createAgentBackendSocket;
 function createAgentBackendSocket({ WebSocketImpl, wsUrl, wsOrigin, headers, }) {
     if (!WebSocketImpl) {
         throw new Error('Agent SDK backend socket requires WebSocketImpl');
@@ -16,5 +17,4 @@ function createAgentBackendSocket({ WebSocketImpl, wsUrl, wsOrigin, headers, }) 
         headers: headers || {},
     });
 }
-exports.createAgentBackendSocket = createAgentBackendSocket;
 exports.createWindieSdkBackendSocket = createAgentBackendSocket;